Enable AVX 2 if the codegen target-feature is enabled.
AVX 2 code generation without `target-feature=avx2` enabled is still not
properly functioning as of Rust 1.30.1. However, when
`target-feature=avx2` is enabled, the correct code is generated. We'll
just require `target-feature=avx2` for that bit of code, as calling AVX
intrinsics via a non-AVX ABI was bound to incur a performance penalty
anyway.